requirements.txt
absl-py==0.9.0
astunparse==1.6.3
awsebcli==3.18.1
botocore==1.15.49
cachetools==4.1.1
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
future==0.16.0
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.7
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.1
pathspec==0.5.9
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pypiwin32==223
python-dateutil==2.8.0
pywin32==228
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
semantic-version==2.5.0
six==1.11.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
urllib3==1.24.3
wcwidth==0.1.9
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
